+++ lib/core/router/router.dart
@@ -8,6 +8,7 @@ import '../../features/recurrences/presentation/manage_recurrences_page.dart';
 import '../../features/recurrences/presentation/recurrences_list_page.dart';
 import '../../features/settings/presentation/settings_page.dart';
 import '../../features/statistics/presentation/statistics_page.dart';
+import '../../features/members/presentation/manage_members_page.dart';
 import '../services/database_service.dart';
 
 final router = GoRouter(
@@ -53,6 +54,10 @@ final router = GoRouter(
     GoRoute(
       path: '/statistics',
       builder: (context, state) => const StatisticsPage(),
+    ),
+    GoRoute(
+      path: '/members',
+      builder: (context, state) => const ManageMembersPage(),
     ),
   ],
 );