+++ lib/features/transactions/data/transaction_repository_impl.dart
@@ -31,8 +31,7 @@ class TransactionRepositoryImpl implements TransactionRepository {
         .getFullList(
           filter: filter,
           sort: '-date',
-          expand:
-              'account,recurrence', // Expand account to get details if needed
+          expand: 'account,recurrence,member',
         );
 
     return records.map((e) => e.toJson()).toList();
@@ -63,7 +62,7 @@ class TransactionRepositoryImpl implements TransactionRepository {
           filter: filter,
           sort:
               '-date', // Most recent overdue first? Or oldest? Maybe oldest on top to clear them? Let's stick to -date consistent with main list
-          expand: 'account,recurrence',
+          expand: 'account,recurrence,member',
         );
 
     return records.map((e) => e.toJson()).toList();